backend/requirements.txt
fastapi==0.111.1
starlette==0.37.2
uvicorn[standard]==0.30.6
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic-settings==2.4.0
email-validator==2.2.0
python-multipart==0.0.9

pytest==8.3.2
httpx==0.27.0
anyio==4.4.0
sniffio==1.3.1
mongomock==4.1.2
